chore: update index.ts in DI module. (core currently unusable)
